feat: Fix phpstan
